+++ CHANGELOG.md
@@ -94,4 +94,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Deprecated** for soon-to-be removed features
 - **Removed** for now removed features
 - **Fixed** for any bug fixes
-- **Security** in case of vulnerabilities
+- **Security** in case of vulnerabilities# Updated 20251109_123817